🚚 Rename files to kebab-case
